refactor: reorganize client folder
