+++ packages/live-state/src/client/stateful/obj-graph.ts
